codegen/strings: Allow for custom parsers
The "error" out parameter was a conscious decision. The other solution
for error logging would've been exposing the offset, so that the parser
could correctly report the error at the position.

But what if one day I implement row/col error logging? Or error logging
with a context pointer, so it could e.g. go into a buffer.

Most errors can be properly logged this way too.

Implements #19