json: enable JSMN_STRICT
It was disabled before, because JSMN was not willing to parse simple
literals if it was enabled. This was only needed for tests, and I found
out that putting a space after the literals makes JSMN happy again, so
there was little change needed.

Now we will avoid the following invalid JSON cases:
{ 1: 2}
{ "wat": 1 2 3}
{"token?": TOKEN }

Some invalid JSON tests had to be changed, and while doing that, I found
an interesting pitfall. While STRICT will catch "INF" and "NAN", it does
not catch "nan", because it begins with a lowercase "n", and is allowed
to allow "null".